Agrego avances en funcionalidad detallepedido y pedido

+++ Dominio/DetallePedido.cs
@@ -10,7 +10,7 @@ namespace Dominio
     {
         public int Id { get; set; }
         public int Cantidad { get; set; }
-        public float PrecioUnitario { get; set; }
+        public decimal PrecioUnitario { get; set; }
         public Insumo Insumo { get; set; }
     }
 }
